When you select the next_process from the ready_queue, you do not remove it from the ready_queue. You just copy it and leave the original in the queue. You need to remove the selected process from the ready_queue.
In your function, you first remove the completed process (index 0) and then find the next process based on the highest priority. However, if the process at index 0 was also the highest priority process, and there is no other process with the same priority, you will not be able to find it anymore, because it has already been removed.