Here's how the function works:
If there is no currently running process (indicated by a process ID of 0), then the new process is scheduled to run. Its execution start time is set to the current timestamp and its execution end time is set to the sum of the current timestamp and the smaller of the total burst time and the time quantum.
If there is a currently running process, the new process is simply added to the ready queue and the current process continues to run. The start time and end time of the new process are set to 0 as it hasn't started executing yet.
Please note that the ready queue is an array of PCBs and the queue_cnt variable is a pointer to an integer that keeps track of the number of items in the ready queue. When a process is added to the ready queue, the count is incremented.