This function begins by iterating through the memory map to find the block with the smallest size that is large enough to accommodate the request (best_fit_block).
If there is no suitable block, the function returns a NULLBLOCK.

If a block of exact size is found, the function updates this block's process_id to allocate it, updates the memory map, and returns the updated block.

If the chosen block is larger than the requested size, the block is split into two: the best_fit_block (which is now allocated) and a new free block (new_free_block). Both are updated in the memory map, the map_cnt is increased, and the best_fit_block is returned.