This function works by looping through each page reference in the reference string. For each reference, it checks if the page is already loaded in the page table (i.e., a "hit"). If it is not (a "miss"), it increases the page fault count.
When a page fault occurs, the function checks if there is a free frame in the frame pool. If there is, it loads the page into the free frame. If there isn't a free frame, it needs to replace a page. Because this function implements the FIFO page replacement algorithm, it selects the oldest page (i.e., the page with the earliest arrival timestamp) and replaces it with the new page.