The function is currently iterating over the page table entries and checking if the frame_number matches the current reference. This would only work if the page table entries were directly mapping the page numbers to frames, but this isn't necessarily the case. The frame_number of a page table entry is just an available frame number taken from the frame pool, it's not related to the index of the page table or the page number.
Instead, when a reference to a page is made, we should check the page table entry that corresponds to that page number, and see if the is_valid bit is true. If it is, then the page is in memory. If it isn't, then a page fault occurs and we need to allocate a frame to the page, either from the frame pool if there are any available, or by replacing an existing page if the frame pool is empty.