It seems that the error lies in the logic used to increment page faults in count_page_faults_fifo. The current condition checks if the frame number returned by process_page_access_fifo differs from the frame number currently stored in the page table. However, this may not correctly capture all page faults, as a page fault occurs when a page is not currently in memory and needs to be loaded, whether the frame number changes or not.
A better way to detect a page fault in process_page_access_fifo could be to check if the page was invalid before the function call. If the page was invalid before the function call and is valid after the call, this means that the page was not in memory and had to be loaded, which is a page fault.