In this revised function, process_page_access_lru is called to process each page access in the reference string. The timestamp passed to process_page_access_lru is i + 1, which corresponds to the position of the page number in the reference string (simulating a timestamp).
The process_page_access_lru function should return -1 if a page fault occurs (i.e., if the page is not already in memory and there are no free frames available). If a page fault has occurred, the count of page faults (faults) is incremented. The total number of page faults is returned at the end.