Fixed FIFO Page Replacement Algorithm
This commit includes fixes to the FIFO page replacement algorithm implemented in the `process_page_access_fifo` function. The changes are:

1. Fixed the frame replacement logic in `process_page_access_fifo`. Previously, the function was incorrectly setting the `arrival_timestamp` for the replaced page to `-1`. This commit changes this to set the `arrival_timestamp` to `current_timestamp` which correctly represents the time of the new page arrival.

2. The function now correctly sets the `reference_count` of the replaced page to `0`. Previously, it was being set to `-1` which does not correctly represent the number of references to a page.

3. In the `count_page_faults_fifo` function, we now correctly increment the `faults` counter when a frame number of `-1` is returned by `process_page_access_fifo`. This represents a page fault. Previously, the function was not checking for a return value of `-1` and hence was not correctly counting page faults.

These changes should result in the correct implementation of the FIFO page replacement algorithm and correctly count the number of page faults using this algorithm.

Further testing is needed to confirm the correct operation of these functions, especially for edge cases and performance in high-load situations.